tests/requirements.txt
requests==2.31.0
dnspython==2.4.2
pysnmp==4.4.12
redis==5.0.1
paramiko==3.4.0 